Fixed check output
Prepend service name and state to the check output message, as required
by the guidelines.

Introduces 'NagiosCheck.service', which should be set to the (short) name
of the check.

Introduces 'NagiosCheck.state_sep' for the separation between state and
message. Defaults to ':' per the guidelines, but can be overriden with
something more common like ' -'.